requirements.txt
Django==2.0.2
djangorestframework==3.7.7
gunicorn==19.7.1
mysqlclient==1.3.12
pytz==2018.3
